java/airflow/requirements.txt
apache-airflow-providers-postgres==5.10.0
soda-core-postgres==3.3.3
dbt-postgres==1.7.4
acryl-datahub==0.12.1
